% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcal.R
\name{bcal}
\alias{bcal}
\title{Calibration of P-Values for Testing Point Null Hypotheses}
\usage{
bcal(p)
}
\arguments{
\item{p}{A numeric vector with values in the [0,1] interval.}
}
\value{
Returns a numeric vector with length as \code{p}.
}
\description{
Calibrate p-values under a robust Bayesian perspective so that they can be interpreted as lower bounds on the odds provided by the data (Bayes factors) in favor of point null hypotheses.
}
\details{
A useful way to calibrate a p-value under a robust Bayesian perspective is by using the bound that is found as the minimum Bayes factor in favour of the null that is obtained by changing priors over large classes of distributions under \eqn{H_{1}{H1}}. This Bayesian calibration is obtained with \deqn{B(p) = -exp(1) p log (p)} when \eqn{p < 1/exp{1}}  \insertCite{pericchiTorres2011}{pcal}  \insertCite{sellke2001}{pcal}
}
\examples{
# Calibration of typical "threshold" p-values:
bcal(c(.1, .05, .01, .005, .001))

# Application: chi-squared goodness-of-fit test,
# lower bound on the Bayes factor in favor of the null:
x <- matrix(c(12, 41, 25, 33), ncol = 2)
bcal(chisq.test(x)[["p.value"]])

}
\references{
\insertAllCited{}
}
\seealso{
{\itemize
\item \code{\link[pcal]{pcal}} For a p-value calibration that returns lower bounds on the posterior probability of the null.
\item \code{\link[pcal]{bfactor_to_prob}} to turn Bayes factors into posterior probabilities.}
}
