% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfactor_interpret.R
\name{bfactor_interpret_2}
\alias{bfactor_interpret_2}
\title{Interpretation of Bayes factors}
\usage{
bfactor_interpret_2(bf)
}
\arguments{
\item{bf}{A numeric vector of non-negative values.}
}
\value{
Returns a character vector with the same \code{length} as \code{bf}.
}
\description{
Quantify the strength of the evidence provided by the data to a model/hypothesis according to a Bayes factor interpretation scale from \insertCite{kass1995;textual}{pcal}.
}
\details{
Bayes factors are a summary of the evidence provided by the data to a model/hypothesis. \insertCite{kass1995;textual}{pcal} remarked that it can be useful to consider twice the natural logarithm of the Bayes factor, which is in the same scale as the familiar deviance and likelihood ratio test statistics. With this in mind, \code{bfactor_interpret_2} takes Bayes factors (\code{bf}) as input and returns the strength of the evidence provided by the data to the  model/hypothesis in the numerator of the Bayes factors (usually the null hypothesis) according to the following scale:\tabular{lll}{
   2*log((bf) \tab bf \tab Evidence \cr
   [-Inf, 0[ \tab [0, 1[ \tab Negative \cr
   [0, 2[ \tab [1, 3[ \tab Weak \cr
   [2, 6[ \tab [3, 20[ \tab Positive \cr
   [6, 10[ \tab [20, 150[ \tab Strong \cr
   [10, +Inf[ \tab [150, +Inf) \tab Very strong \cr
}


\insertCite{jeffreys1961;textual}{pcal} suggested the interpretation of Bayes factors in half-units on the base 10 logarithmic scale (see \code{\link[pcal]{bfactor_interpret}}. Alternatively,   suggested by \insertCite{kass1995;textual}{pcal}:
Two of the original categories from \insertCite{jeffreys1961;textual}{pcal} were pooled together for simplification and a new category for the case when the data provides more support to the alternative model/hypothesis was added.

When comparing results with those from standard likelihood ratio tests, it is convenient to put the null hypothesis in the denominator of the Bayes factor so that \code{bfactor_interpret} returns the strength of the evidence against the null hypothesis. In that case, and if \code{bf} was obtained with the null hypothesis on the numerator, one can use 1/\code{bf} instead.
}
\examples{
# Interpret one Bayes factor
bfactor_interpret(1.5)
bfactor_interpret_2(1.5)

# Interpret many Bayes factors
bfactor_interpret(c(0.1, 1.2, 3.5, 13.9, 150))
bfactor_interpret_2(c(0.1, 1.2, 3.5, 13.9, 150))

# Application: chi-squared goodness-of-fit test.
# Strength of the evidence provided by the lower
# bound on the Bayes factor in favor of the null:
x <- matrix(c(12, 15, 14, 15), ncol = 2)
bfactor_interpret(bcal(chisq.test(x)[["p.value"]]))
bfactor_interpret_2(bcal(chisq.test(x)[["p.value"]]))

}
\references{
\insertAllCited{}
}
\seealso{
\itemize{
\item \code{\link[pcal]{bfactor_interpret}} for an alternative interpretation scale suggested by \insertCite{jeffreys1961;textual}{pcal}.
\item \code{\link[pcal]{bfactor_log_interpret}} for the interpretation of the logarithms of Bayes factors.
\item \code{\link[pcal]{bfactor_to_prob}} to turn Bayes factors into posterior probabilities.
\item \code{\link[pcal]{bcal}} for the calculation of lower bounds on Bayes factors.
}
}
