% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{check_log_base}
\alias{check_log_base}
\title{Check if an object is a valid logarithmic base}
\usage{
check_log_base(base)
}
\arguments{
\item{bf}{An arbitrary object.}
}
\value{
\code{check_log_base} does not return any output. There are two possible scenarios:
\itemize{
\item The call is silent if \code{base} is a numeric vector of \code{\link[base]{length}} 1 that is a valid logarithmic base.
\item An informative error message is thrown otherwise.
}
}
\description{
\code{check_log_base} checks if an object is a numeric vector of \code{\link[base]{length}} 1 that is eligible to be used as a logarithmic base. This can be useful to validate inputs in user-defined functions.
}
\details{
\code{check_log_base} conducts a series of tests to check if \code{base} is a valid logarithmic base. Namely, \code{check_log_base} checks if:
\itemize{
\item \code{base} is \code{NULL} or empty.
\item \code{base} is a numeric (atomic) vector of \code{\link[base]{length}} 1.
\item \code{base} is \code{NA} or \code{NaN}.
}
}
\examples{
# Calls that pass silently:
check_log_base(10)
check_log_base(2)
check_log_base(exp(1))
check_log_base(0.5)

# Calls that throw informative error messages:
\dontrun{check_log_base(NULL)}
\dontrun{check_log_base(numeric(0))}
\dontrun{check_log_base(-1)}
\dontrun{check_log_base(0)}
\dontrun{check_log_base(TRUE)}
\dontrun{check_log_base("10")}
\dontrun{check_log_base(factor(10))}
\dontrun{check_log_base(matrix(10))}
\dontrun{check_log_base(list(10))}
\dontrun{check_log_base(NA)}
\dontrun{check_log_base(NaN)}

}
\seealso{
\itemize{
\item \code{\link[pcal]{check_prob}} to check if an object is a numeric vector of valid probability values.
\item \code{\link[pcal]{check_bf}} to check if an object is a numeric vector of valid Bayes factor values.
\item \code{\link[pcal]{check_log_bf}} to check if an object is a numeric vector of valid logarithmic Bayes factor values.
}
}
\keyword{internal}
