% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-functions.R
\name{check_log_bf}
\alias{check_log_bf}
\title{Check if an object is a numeric vector of valid logarithmic Bayes factor values}
\usage{
check_log_bf(bf)
}
\arguments{
\item{bf}{An arbitrary object.}
}
\description{
\code{check_log_bf} checks if an object is a numeric vector of valid logarithmic Bayes factor values. This can be useful to validate inputs and intermediate calculations in user-defined functions or to validate the output of calculations/functions.
}
