% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-functions.R
\name{check_prob}
\alias{check_prob}
\title{Check if an object is a vector of valid probability values}
\usage{
check_prob(p)
}
\arguments{
\item{p}{An arbitrary object.}
}
\value{
\code{check_prob} does not return any output. There are three possible scenarios:
\itemize{
\item The call is silent if \code{p} is a numeric vector of valid probability values and there are no \code{NA} or \code{NaN} values.
\item An informative warning message is given if \code{p} is a numeric vector of valid probability values and there are \code{NA} or \code{NaN} values.
\item An informative error message is thrown if \code{p} is not a numeric vector of valid probability values.
}
}
\description{
\code{check_prob} checks if an object is a numeric vector of valid probability values. This can be useful to validate inputs and intermediate calculations in user-defined functions or to validate the output of calculations/functions.
}
\details{
\code{check_prob} conducts a series of tests to check if \code{p} is a numeric vector of valid probability values. Namely, \code{check_prob} checks if:
\itemize{
\item \code{p} is \code{NULL} or empty.
\item \code{p} is a numeric (atomic) vector.
\item \code{p} has \code{NA} or \code{NaN} values.
\item The values of \code{p} are in the [0, 1] interval.
}
}
\examples{
# Calls that pass silently:
check_prob(0)
check_prob(0.5)
check_prob(1)
check_prob(c(0.1, 0.2, 0.3, 0.4, 0.5))

# Call that throws an informative warning message:
\dontrun{check_prob(c(0.1, 0.2, NA, 0.4, 0.5))}

# Calls that throw informative error messages:
\dontrun{check_prob(NULL)}
\dontrun{check_prob()}
\dontrun{check_prob(TRUE)}
\dontrun{check_prob("0.5")}
\dontrun{check_prob(factor(0.5))}
\dontrun{check_prob(matrix(0.5))}
\dontrun{check_prob(list(0.5))}
\dontrun{check_prob(NA)}
\dontrun{check_prob(NaN)}
\dontrun{check_prob(1.1)}
\dontrun{check_prob(-0.5)}
\dontrun{check_prob(c(-0.9, 0, 0.1, 0.2, 0.3, 0.4, 0.5))}

}
\seealso{
\itemize{
\item \code{\link[pcal]{check_bf}} To check if an object is a numeric vector of valid Bayes factor values.
\item \code{\link[pcal]{check_log_bf}} To check if an object is a numeric vector of valid logarithmic Bayes factor values.
}
}
