% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{check_scale}
\alias{check_scale}
\title{Check if an object is a string of characters representing an eligible Bayes factor interpretation scale}
\usage{
check_scale(scale)
}
\arguments{
\item{scale}{An arbitrary object.}
}
\value{
\code{check_scale} does not return any output. There are two possible scenarios:
\itemize{
\item The call is silent if \code{base} is a string of characters representing one of the Bayes factor interpretation scales available in the \code{pcal} package.
\item An informative error message is thrown otherwise.
}
}
\description{
\code{check_scale} checks if an object is a string of characters representing one of the Bayes factor interpretation scales available in the \code{pcal} package. This can be useful to validate inputs in user-defined functions.
}
\details{
\code{check_scale} conducts a series of tests to check if \code{scale} is a string of characters representing one of the Bayes factor interpretation scales available in the \code{pcal} package. Namely, \code{check_scale} checks if:
\itemize{
\item \code{scale} is \code{NULL} or empty.
\item \code{scale} is a character (atomic) vector of \code{\link[base]{length}} 1 specifying either "Jeffreys" or "Kass-Raftery" (not case sensitive).
\item \code{scale} is \code{NA} or \code{NaN}.
}
}
\examples{
# Calls that pass silently:
check_scale("Jeffreys")
check_scale("jeffreys")
check_scale("kass-raftery")
check_scale("Kass-Raftery")

# Calls that throw informative error messages:
\dontrun{check_scale(NULL)}
\dontrun{check_scale(10)}
\dontrun{check_scale("Bayes")}
\dontrun{check_scale("jeff")}
\dontrun{check_scale("kassraftery")}
\dontrun{check_scale(c("jeffreys", "kass-raftery"))}

}
\seealso{
\itemize{
\item \code{\link[pcal]{check_prob}} to check if an object is a numeric vector of valid probability values.
\item \code{\link[pcal]{check_bf}} to check if an object is a numeric vector of valid Bayes factor values.
\item \code{\link[pcal]{check_log_bf}} to check if an object is a numeric vector of valid logarithmic Bayes factor values.
\item \code{\link[pcal]{check_log_base}} to check if an object is a numeric vector of \code{\link[base]{length}} 1 representing a valid logarithmic base.
}
}
\keyword{internal}
