% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcal.R
\docType{package}
\name{pcal-package}
\alias{pcal}
\alias{pcal-package}
\title{pcal: Calibration of P-Values for Point Null Hypothesis Testing}
\description{
Calibrate p-values under a robust perspective using the methods developed by \insertCite{sellke2001;textual}{pcal} and obtain measures of the evidence provided by the data in favor of point null hypotheses which are safer and more straightforward to interpret.

For a complete list of functions see \code{library(help = "pcal")}. For tutorials and examples see the \href{https://github.com/pedro-teles-fonseca/pcal/blob/master/README.md}{README}.
}
\references{
\insertAllCited{}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pedro-teles-fonseca.github.io/pcal/}
  \item \url{https://github.com/pedro-teles-fonseca/pcal}
  \item Report bugs at \url{https://github.com/pedro-teles-fonseca/pcal/issues}
}

}
\author{
\strong{Maintainer}: Pedro Fonseca \email{pedro.teles.fonseca@outlook.com} (\href{https://orcid.org/0000-0002-5708-081X}{ORCID})

Other contributors:
\itemize{
  \item Rui Paulo (\href{https://orcid.org/0000-0002-4802-527X}{ORCID}) [contributor, thesis advisor]
  \item FCT [funder]
}

}
\keyword{internal}
