% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcal.R
\name{pcal}
\alias{pcal}
\title{Lower Bounds on the Posterior Probabilities of Point Null Hypotheses}
\usage{
pcal(p, prior_prob = 0.5)
}
\arguments{
\item{p}{A numeric vector with values in the [0,1] interval.}

\item{prior_prob}{A numeric vector with values in the [0,1] interval.}
}
\value{
Returns a numeric vector with length as \code{p}.
}
\description{
Calibrate p-values under a robust Bayesian perspective so that they can be interpreted as either lower bounds on the probability of a type I error or lower bounds on the posterior probability of the null hypothesis.
}
\details{

}
\examples{
# Calibration of typical "threshold" p-values:
pcal(c(.1, .05, .01, .005, .001))

# Application: chi-squared goodness-of-fit test,
# lower bound on the posterior probability of the null:
x <- matrix(c(12, 41, 25, 33), ncol = 2)
pcal(chisq.test(x)[["p.value"]])

}
\references{
\insertAllCited{}
}
\seealso{
{\link[pcal]{bcal}} for a p-value calibration that returns lower bounds on the odds provided by the data (Bayes factors) in favor of point null hypotheses.
}
