% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcal.R
\name{bcal}
\alias{bcal}
\title{Lower bounds on Bayes for Testing Point Null Hypotheses}
\usage{
bcal(p)
}
\arguments{
\item{p}{A numeric vector with values in the [0,1] interval.}
}
\value{
Returns a numeric vector with length as \code{p}.
}
\description{
Calibrate p-values under a robust Bayesian perspective so that they can be interpreted as lower bounds on the odds provided by the data (Bayes factors) in favor of point null hypotheses.
}
\details{
To obtain a robust measure of the evidence provided by the data in favor of a point the null hypothesis we can derive the lower bound on the odds of the data (or Bayes factor) that is found when changing the prior distribution of the parameters of interest under the alternative hypothesis over a wide class of distributions. \code{bcal} approximates such a lower bound using the p-value calibration developed in \insertCite{sellke2001;textual}{pcal}: \deqn{B(p) = -exp(1) p log (p)} for \eqn{p < 1/exp(1)}, where \code{p} is a  p-value on a classical test statistic, and \code{bcal(p) = 1} otherwise. \insertCite{sellke2001;textual}{pcal} note that a scenario in which they definitely recommend use of this calibration is then investigating fit to the null model, with no explicit alternative in mind. \insertCite{pericchiTorres2011;textual}{pcal} warn, however, that despite useful the p-values calibrations developed in \insertCite{sellke2001;textual}{pcal} do not depend on sample size and hence the lower bounds obtained with large sample sizes may be conservative
}
\examples{
# Calibration of typical "threshold" p-values:
bcal(c(.1, .05, .01, .005, .001))

# Application: chi-squared goodness-of-fit test,
# lower bound on the Bayes factor in favor of the null:
x <- matrix(c(12, 41, 25, 33), ncol = 2)
bcal(chisq.test(x)[["p.value"]])

}
\references{
\insertAllCited{}
}
\seealso{
{\itemize
\item \code{\link[pcal]{pcal}} for a p-value calibration that returns lower bounds on the posterior probability of the null.
\item \code{\link[pcal]{bfactor_to_prob}} to turn Bayes factors into posterior probabilities.}
}
