% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfactor_interpret.R
\name{bfactor_interpret}
\alias{bfactor_interpret}
\title{Interpretation of Bayes factors}
\usage{
bfactor_interpret(bf, scale = "jeffreys")
}
\arguments{
\item{bf}{A numeric vector of non-negative values.}

\item{scale}{A string of characters specifying either "jeffreys" or "kass-raftery". Not case sensitive.}
}
\value{
Returns a character vector with the same {\link[base]{length}} as \code{bf}.
}
\description{
Quantify the strength of the evidence provided by the data to a model/hypothesis according to Bayes factor interpretation scales suggested by \insertCite{jeffreys1961;textual}{pcal} and \insertCite{kass1995;textual}{pcal}.
}
\details{
Bayes factors are a summary of the evidence provided by the data to a model/hypothesis. \insertCite{jeffreys1961;textual}{pcal} suggested the interpretation of Bayes factors in half-units on the base 10 logarithmic scale, as indicated in the following table:\tabular{lll}{
   \strong{log10(Bayes factor)} \tab \strong{Bayes factor} \tab \strong{Evidence} \cr
   [-Inf, 0[ \tab [0, 1[ \tab Negative \cr
   [0, 0.5[ \tab [1, 3.2[ \tab Weak \cr
   [0.5, 1[ \tab [3.2, 10[ \tab Substantial \cr
   [1, 1.5[ \tab [10, 32[ \tab Strong \cr
   [1.5, 2[ \tab [32, 100[ \tab Very Strong \cr
   [2, +Inf[ \tab [100, +Inf[ \tab Decisive \cr
}


By default, \code{bfactor_interpret} takes Bayes factors as input and returns the strength of the evidence in favor of the model/hypothesis in the numerator of the Bayes factors (usually the null hypothesis) according to the aforementioned table.

Alternatively, and because it can be useful to consider twice the natural logarithm of the Bayes factor (which is in the same scale as the familiar deviance and likelihood ratio test statistics), \insertCite{kass1995;textual}{pcal} suggested the following scale:\tabular{lll}{
   \strong{2*log(Bayes factor)} \tab \strong{Bayes factor} \tab \strong{Evidence} \cr
   [-Inf, 0[ \tab [0, 1[ \tab Negative \cr
   [0, 2[ \tab [1, 3[ \tab Weak \cr
   [2, 6[ \tab [3, 20[ \tab Positive \cr
   [6, 10[ \tab [20, 150[ \tab Strong \cr
   [10, +Inf[ \tab [150, +Inf[ \tab Very Strong \cr
}


To interpret Bayes factors according to the latter table use \code{scale = "kass-raftery"}.

When comparing Bayes factors with results of standard likelihood ratio tests it is convenient to put the null hypothesis in the denominator of the Bayes factor so that \code{bfactor_interpret} returns the strength of the evidence against the null hypothesis. If \code{bf} was obtained with the null hypothesis on the numerator, one can use \code{bfactor_interpret(1/bf)} or \code{bfactor_interpret(1/bf, scale = "kass-raftery")} to obtain the strength of the evidence against the null hypothesis.
}
\examples{
# Interpretation of one Bayes factor
bfactor_interpret(4)
bfactor_interpret(4, scale = "jeffreys")
bfactor_interpret(4, scale = "kass-raftery")

# Interpretation of more than one Bayes factors
bfactor_interpret(c(0.1, 1.2, 3.5, 13.9, 150))
bfactor_interpret(c(0.1, 1.2, 3.5, 13.9, 150), scale = "jeffreys")
bfactor_interpret(c(0.1, 1.2, 3.5, 13.9, 150), scale = "kass-raftery")

# Application: chi-squared goodness-of-fit test.
# Strength of the evidence provided by the lower
# bound of the Bayes factor in favor of the null hypothesis:
data <- matrix(c(18, 12, 10, 12, 10, 23), ncol = 2)
bfactor_interpret(bcal(chisq.test(data)[["p.value"]]))

}
\references{
\insertAllCited{}
}
\seealso{
\itemize{
\item \code{\link[pcal]{bfactor_log_interpret}} for the interpretation of the logarithms of Bayes factors.
\item \code{\link[pcal]{bfactor_to_prob}} to turn Bayes factors into posterior probabilities.
\item \code{\link[pcal]{bcal}} for a p-value calibration that returns lower bounds on Bayes factors in favor of point null hypotheses.
}
}
