% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfactor_interpret.R
\name{bfactor_interpret}
\alias{bfactor_interpret}
\title{Interpretation of Bayes factors}
\usage{
bfactor_interpret(bf)
}
\arguments{
\item{bf}{A numeric vector of non-negative values.}
}
\value{
Returns a character vector with the same \code{length} as \code{bf}.
}
\description{
Quantify the strength of the evidence provided by the data to a model/hypothesis according to the Bayes factor interpretation scale from \insertCite{kass1995;textual}{pcal}.
}
\details{
Bayes factors are a summary of the evidence provided by the data in favor of a model/hypothesis. \code{bfactor_interpret} takes Bayes factors (\code{bf}) as input and returns the strength of the evidence provided by the data in favor of the  model/hypothesis in the numerator of the Bayes factors (usually the null hypothesis) according to the scale suggested by \insertCite{kass1995;textual}{pcal}. The original scale was augmented with an additional category for the case when the data provides more support to the alternative hypothesis:\tabular{ll}{
   Bayes factor \tab Evidence \cr
   (-Inf, 1[ \tab Negative \cr
   [1, 3.2[ \tab Weak \cr
   [3.2, 10[ \tab Substantial \cr
   [10, 100[ \tab Strong \cr
   [100, +Inf) \tab Decisive \cr
}


To obtain the strength of the evidence against the null hypothesis use 1/\code{bf} instead of \code{bf}, which is equivalent to putting the null hypothesis in the denominator of the Bayes factor. This is useful for comparing results with standard likelihood ratio tests.
}
\examples{
# Interpret one Bayes factor
bfactor_interpret(1.5)

# Interpret many Bayes factors
bfactor_interpret(c(0.1, 1.2, 3.5, 13.9, 150))

# Application: chi-squared goodness-of-fit test,
# strength of the evidence provided by the lower
# bound on the Bayes factor in favor of the null:
x <- matrix(c(12, 15, 14, 15), ncol = 2)
bfactor_interpret(bcal(chisq.test(x)[["p.value"]]))

}
\references{
\insertAllCited{}
}
\seealso{
\itemize{
\item \code{\link[pcal]{bfactor_log_interpret}} to interpret the logarithms of Bayes factors.
\item \code{\link[pcal]{bcal}} for the calculation of Bayes factors.
\item \code{\link[pcal]{bfactor_to_prob}} to turn Bayes factors into posterior probabilities.
}
}
