% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfactor_to_prob.R
\name{bfactor_to_prob}
\alias{bfactor_to_prob}
\title{Posterior Probabilities of Point Null Hypotheses}
\usage{
bfactor_to_prob(bf, prior_prob = 0.5)
}
\arguments{
\item{bf}{A numeric vector of non-negative values.}

\item{prior_prob}{A numeric vector with values in the [0,1] interval. If \code{length(bf)} = 1 then \code{prior_prob} can be of any positive length, but if \code{length(bf)} > 1 then \code{length(prior_prob)} must be either 1 or equal to \code{length(bf)}.}
}
\value{
If \code{length(bf)} > 1 then \code{bfactor_to_prob} returns a numeric vector with the same \code{length} as \code{bf}. Otherwise it returns a numeric vector with the same \code{length} as \code{prior_prob}.
}
\description{
Update the prior probabilities of point null hypotheses to posterior probabilities using Bayes factors.
}
\details{
\code{bfactor_to_prob} computes the posterior probabilities of point null hypotheses using the following relation from \insertCite{bergerDelampady1987;textual}{pcal}: \deqn{Prob(H_0|data) = \left(1 + \frac{1 - {null\_prob}}{null\_prob} \times \frac{1}{bf}\right)^{-1}}{Prob(null | data) = (1 + (1 - prior_prob) / prior_prob * (1 / bf)) ^(-1)} where \code{bf} is a Bayes factor in favor of a point null hypothesis with prior probability \code{prior_prob}. The alternative hypothesis has prior probability 1-\code{prior_prob}. \code{prior_prob} can only be of \code{length} 1, in which case it will be recycled (if necessary) and hence every element of \code{bf} will combined with the same prior probability, or of \code{length} equal to \code{length(bf)}, in which case each prior probability in \code{null.par} will be updated using the corresponding element of \code{bf}. The optional argument \code{prior_prob} is set to 0.5 by default, implying prior equiprobability of hypotheses.
}
\examples{
# With a Bayes factor that is indifferent between the null and the alternative:
bfactor_to_prob(1)

# Same as above but the null has high prior probability:
bfactor_to_prob(1, .99)

# Posterior probability of the null as a function of different prior probabilities:
bfactor_to_prob(1, seq(.5, 1, .1))

# With Bayes factors that favor the null hypothesis:
round(bfactor_to_prob(seq(2, 50, 2.5)), 3)

# Same as above but the null has low prior probability:
round(bfactor_to_prob(seq(2, 50, 2.5), prior_prob = .01), 3)

# Posterior probabilities obtained with Bayes factors that favor the alternative:
round(bfactor_to_prob(seq(0, 1, .05)), 3)

# Same as above but the null has high prior probability:
round(bfactor_to_prob(seq(0, 1, .05), prior_prob = .99), 3)

# Application: chi-squared goodness-of-fit test,
# lower bound on the posterior probability of the null:
x <- matrix(c(12, 41, 25, 33), ncol = 2)
bfactor_to_prob(bcal(chisq.test(x)[["p.value"]]), prior_prob = .9)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[pcal]{bcal}} for the calculation of Bayes factors.
}
