% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfactor_to_prob.R
\name{bfactor_to_prob}
\alias{bfactor_to_prob}
\title{Posterior Probabilities of Precise Null Hypotheses in terms of Bayes Factors}
\usage{
bfactor_to_prob(bf, null_prob = 0.5)
}
\arguments{
\item{bf}{A non-negative numeric vector or matrix.}

\item{null_prob}{A numeric vector of length 1. Must be between 0 and 1 (inclusive).}
}
\value{
\code{bfactor_to_prob} returns:
\itemize{
  \item A numeric vector with the same \code{length} as \code{bf} if \code{bf} is a vector.
  \item A numeric matrix with the same \code{dim} as \code{bf} if \code{bf} is a matrix.}
}
\description{
Update the prior probability of a precise null hypothesis to the posterior probability using a Bayes factor.
}
\details{
\code{bfactor_to_prob} takes as input a Bayes factor in favor of the null hypothesis (\code{bf}) and the prior probability of the null hypothesis (\code{null_prob}). It then computes the posterior probability of the null hypothesis using the following relation from \insertCite{bergerSelke1987;textual}{pcal}: \deqn{(1 + (1 - null_prob) / null_prob * (1 / bf)) ^(-1)} The argument \code{null_prob} is optional and set to 0.5 by default, implying equiprobability of events.
}
\examples{
# With Bayes factors that favor the null hypothesis:
round(bfactor_to_prob(seq(2, 50, 2.5)), 3)

# When the Bayes factor is indifferent between null and alternative:
bfactor_to_prob(1)

# With Bayes factors that favor the alternative hypothesis:
round(bfactor_to_prob(seq(0, 1, .05)), 3)

# Application: chi-squared goodness of fit test
x <- matrix(c(12, 41, 25, 33), ncol = 2)
bfactor_to_prob(bcal(chisq.test(x)$p.value))

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{bcal}} for the calculation of Bayes factors.
}
