% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-functions.R
\name{check_log_bf}
\alias{check_log_bf}
\title{Check if an object is a numeric vector of valid logarithmic Bayes factor values}
\usage{
check_log_bf(bf)
}
\arguments{
\item{bf}{An arbitrary object.}
}
\value{
\code{check_log_bf} does not return any output. There are three possible scenarios:
\itemize{
\item The call is silent if \code{bf} is a numeric vector of valid logarithmic Bayes factor values and there are no \code{NA} or \code{NaN} values.
\item An informative warning message is given if \code{bf} is a numeric vector of valid logarithmic Bayes factor values and there are \code{NA} or \code{NaN} values.
\item An informative error message is thrown if \code{bf} is not a numeric vector of valid logarithmic Bayes factor values.
}
}
\description{
\code{check_log_bf} checks if an object is a numeric vector of valid logarithmic Bayes factor values. This can be useful to validate inputs or intermediate calculations in user-defined functions or to validate the output of calculations/functions.
}
\details{
\code{check_log_bf} conducts a series of tests to check if \code{bf} is a numeric vector of valid logarithmic Bayes factor values. Namely, \code{check_bf} checks if:
\itemize{
\item \code{bf} is \code{NULL} or empty.
\item \code{bf} is a numeric (atomic) vector.
\item \code{bf} has \code{NA} or \code{NaN} values.
}
}
\examples{
# Calls that pass silently:
check_log_bf(0)
check_log_bf(0.5)
check_log_bf(100)
check_log_bf(c(0, 0.5, 1, 10, 50, 100))
check_log_bf(-0.5)
check_log_bf(-5)
check_log_bf(c(-0.9, 0, 0.1, 0.2, 0.3, 0.4, 0.5))

# Call that throws an informative warning message:
\dontrun{check_log_bf(c(0.1, 2, NA, 40, 0.5))}

# Calls that throw informative error messages:
\dontrun{check_log_bf(NULL)}
\dontrun{check_log_bf()}
\dontrun{check_log_bf(TRUE)}
\dontrun{check_log_bf("0.5")}
\dontrun{check_log_bf(factor(0.5))}
\dontrun{check_log_bf(matrix(0.5))}
\dontrun{check_log_bf(list(0.5))}
\dontrun{check_log_bf(NA)}
\dontrun{check_log_bf(NaN)}

}
\seealso{
\itemize{
\item \code{\link[pcal]{check_bf}} to check if an object is a numeric vector of valid Bayes factor values.
\item \code{\link[pcal]{check_prob}} to check if an object is a numeric vector of valid probability values.
}
}
