% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcal.R
\name{pcal}
\alias{pcal}
\title{Calibration of P-Values for Testing Precise Null Hypotheses}
\usage{
pcal(p, pi_null = 0.5)
}
\arguments{
\item{pi_null}{A numeric of length one value between zero and one.}
}
\description{
\code{pcal} calibrates p-values under a robust Bayesian perspective so that they can be interpreted in either a Bayesian or a frequentist way.
}
\details{
A useful way to calibrate a p-value under a robust Bayesian perspective is by using the bound that is found as the minimum posterior probability of \eqn{H_0} that is obtained by changing priors over large classes of priors under \eqn{H_{1}{H1}}.
}
\examples{
round(pcal(c(.1, .05, .01)), 3)
df <- read.csv("https://goo.gl/j6lRXD")
table(df$treatment, df$improvement)

pcal(chisq.test(df$treatment, df$improvement, correct=FALSE)[["p.value"]])

}
\seealso{
\code{\link{bcal}}
}
